strip ResponseWrapper before stability calc
If the supplied waveform for Fisher calculation has a ResponseWrapper, the waveform calculation is significantly slower. While the wrapper is important before calculating the final Fisher, it is not important for stable surface calculations. In the FisherStability function, I now strip the waveform of the ResponseWrapper before consistently generating the Fisher elements to check for stability. This significantly improved GPU-based runtimes (factor 3x faster on my A100s).